// src/types.ts
// Интерфейс для представления курса валюты в конкретный момент времени.
export interface CurrencyRate {
  // Код валюты, например, "USD", "EUR", "BTC" и т.д.
  code: string;
  // Значение курса валюты.
  rate: number;
  // Временная метка (timestamp) курса валюты, обычно в формате Unix-времени.
  timestamp: number;
}

// Интерфейс для представления исторических данных курса валюты.
export interface HistoricalData {
  // Дата, к которой относится значение курса, например, "2023-10-05".
  date: string;
  // Значение курса валюты в указанную дату.
  rate: number;
}

// Интерфейс для представления пары валют, используемых в конвертации или сравнении.
export interface CurrencyPair {
  // Базовая валюта (от которой производится перевод).
  from: string;
  // Целевая валюта (в которую производится перевод).
  to: string;
}

// Тип, который определяет режим валют: "fiat" для фиатных валют или "crypto" для криптовалют.
export type CurrencyMode = 'fiat' | 'crypto';

// Интерфейс для представления расширенной информации о валюте.
export interface CurrencyData {
  // Код валюты.
  code: string;
  // Название валюты, например, "US Dollar" или "Bitcoin".
  name: string;
  // Текущий курс валюты.
  rate: number;
  // Изменение курса валюты (например, процентное изменение за определённый период).
  change: number;
}